Add option to not seed lifetime profiler with live allocations
The lifetime profiler is used for pprof/heap to examine allocations in a
fixed time window. But, it is seeded with live sampled objects (those
that have not been deleted) when it is created. This obfuscates the
output with existing allocations. We should have a way to not seed the
lifetime allocator with live allocations.